refactor: change default conversion function for correctness witnesses in MEA
Proofs are not compared at the moment, but this may be needed in the
future. Passing the full trace to the cache makes little sense, since
currently proofs are represented as conditions and invariants (and
invariants may be absent). Therefore, conditions are the best candidate
for cached results to enable comparison of different proof versions.